Extracted price settings validation methods to reuse them in the the bulk price setting view model
